[GH-89] workload excel 1 work again
